move function add tests and prevent & geting replaced multiple times
