+++ app/gui/src/project-view/components/visualizations/TableVisualization/tableVizUtils.ts
@@ -1,3 +1,7 @@
+import { ICellRendererParams } from "ag-grid-enterprise"
+import { TextFormatOptions } from "../TableVisualization.vue"
+import { LINKABLE_URL_REGEX } from "@/util/link"
+
 export const getCellValueType = (item: string) => {
   switch (true) {
     case isInteger(item):
@@ -34,4 +38,58 @@ export const isDateTime = (item: string) => {
 export const isNumericType = (valueType: string) => {
   const isNumber = ['Integer', 'Float', 'Decimal', 'Byte']
   return isNumber.indexOf(valueType) != -1
+}
+
+const replaceLinksWithTag = (str: string) => {
+  return str.replace(
+    LINKABLE_URL_REGEX,
+    (url: string) => `<a href="${url}" target="_blank" class="link">${url}</a>`,
+  )
+}
+
+export const formatText = (input: string, textFormatterSelected: TextFormatOptions) => {
+  const htmlEscaped = input
+    .replaceAll('<', '&lt;')
+    .replaceAll('>', '&gt;')
+
+  if (textFormatterSelected === 'off') {
+    const replaceLinks = replaceLinksWithTag(htmlEscaped)
+    return replaceLinks.replace(/^\s+|\s+$/g, '&nbsp;')
+  }
+
+  const partialMappings = {
+    '\r': '<span style="color: #df8800">␍</span> <br>',
+    '\n': '<span style="color: #df8800;">␊</span> <br>',
+    '\t': '<span style="color: #df8800; white-space: break-spaces;">&#8594;  |</span>',
+  }
+  const fullMappings = {
+    '\r': '<span style="color: #df8800">␍</span> <br>',
+    '\n': '<span style="color: #df8800">␊</span> <br>',
+    '\t': '<span style="color: #df8800; white-space: break-spaces;">&#8594;  |</span>',
+  }
+
+  const replaceSpaces =
+    textFormatterSelected === 'full' ?
+      htmlEscaped.replaceAll(' ', '<span style="color: #df8800">&#183;</span>')
+    : htmlEscaped.replace(/ \s+|^ +| +$/g, function (match: string) {
+        return `<span style="color: #df8800">${match.replaceAll(' ', '&#183;')}</span>`
+      })
+
+  const replaceLinks = replaceLinksWithTag(replaceSpaces)
+
+  const replaceReturns = replaceLinks.replace(
+    /\r\n/g,
+    '<span style="color: #df8800">␍␊</span> <br>',
+  )
+
+  const renderOtherWhitespace = (match: string) => {
+    return textFormatterSelected === 'full' && match != ' ' ?
+        '<span style="color: #df8800">&#9744;</span>'
+      : match
+  }
+  const newString = replaceReturns.replace(/[\s]/g, function (match: string) {
+    const mapping = textFormatterSelected.value === 'full' ? fullMappings : partialMappings
+    return mapping[match as keyof typeof mapping] || renderOtherWhitespace(match)
+  })
+  return `<span > ${newString} <span>`
 }
